+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 - ✨(domains) domain accesses update API #423
 - ✨(backend) domain accesses create API #428
 - 🥅(frontend) catch new errors on mailbox creation #392
+- ✨(api) domain accesses delete API #433
 
 ### Fixed